refactor following review
