+++ cishouseholds/pipeline/high_level_transformations.py
@@ -127,7 +127,9 @@ from cishouseholds.impute import impute_latest_date_flag
 from cishouseholds.impute import impute_outside_uk_columns
 from cishouseholds.impute import impute_visit_datetime
 from cishouseholds.impute import merge_previous_imputed_values
+from cishouseholds.pipeline.config import get_config
 from cishouseholds.pipeline.generate_outputs import generate_stratified_sample
+from cishouseholds.pipeline.input_file_processing import extract_lookup_csv
 from cishouseholds.pipeline.mapping import column_name_maps
 from cishouseholds.pipeline.regex_patterns import at_school_pattern
 from cishouseholds.pipeline.regex_patterns import at_university_pattern
@@ -145,6 +147,12 @@ from cishouseholds.pipeline.regex_testing import priority_map
 from cishouseholds.pipeline.regex_testing import roles_map
 from cishouseholds.pipeline.regex_testing import social_care_classification
 from cishouseholds.pipeline.timestamp_map import cis_digital_datetime_map
+from cishouseholds.pipeline.translate import backup_and_replace_translation_lookup_df
+from cishouseholds.pipeline.translate import export_responses_to_be_translated_to_translation_directory
+from cishouseholds.pipeline.translate import get_new_translations_from_completed_translations_directory
+from cishouseholds.pipeline.translate import get_welsh_responses_to_be_translated
+from cishouseholds.pipeline.translate import translate_welsh_fixed_text_responses_digital
+from cishouseholds.pipeline.translate import translate_welsh_free_text_responses_digital
 from cishouseholds.pipeline.validation_schema import validation_schemas  # noqa: F401
 from cishouseholds.pyspark_utils import get_or_create_spark_session
 from cishouseholds.validate_class import SparkValidate
@@ -383,6 +391,85 @@ def transform_survey_responses_version_1_delta(df: DataFrame) -> DataFrame:
     df = apply_value_map_multiple_columns(df, column_editing_map)
     df = clean_barcode(df=df, barcode_column="swab_sample_barcode", edited_column="swab_sample_barcode_edited_flag")
     df = clean_barcode(df=df, barcode_column="blood_sample_barcode", edited_column="blood_sample_barcode_edited_flag")
+    return df
+
+
+def translate_welsh_survey_responses_version_digital(df: DataFrame) -> DataFrame:
+    """
+    High-level function that manages the translation of both fixed-text and free-text Welsh responses in the
+    CIS Digital response extract.
+
+    Loads values from config to determine if the translation_project_workflow_enabled flag is true, if the
+    translation_project_workflow_enabled flag is true then it works through the translation project workflow to:
+    1. check for new completed free-text translations
+    2. if there are new completed free-text translations then backup and update the translations_lookup_df
+    3. update the df from the translations_lookup_df
+    4. filter the df to get free-text responses to be translated
+    5. export free-text responses to be translated
+
+    If the translation_project_workflow_enabled flag is false then it will just carry out step 3. above.
+    After this check it will update the fixed-text responses in the df.
+
+    This function requires that the config contains a valid translation_lookup_path in the storage dictionary of
+    the pipeline_config file.
+
+    Args:
+        df (DataFrame): df containing free- or fixed-text responses to translate
+
+    Returns:
+        DataFrame: df incorporating any available translations to free- or fixed-text responses
+    """
+    translation_settings = get_config().get("translation", {"inactive": "inactive"})
+    storage_settings = get_config().get("storage", {"inactive": "inactive"})
+    translation_lookup_path = storage_settings.get("translation_lookup_path", "inactive")
+
+    translation_settings_in_pipeline_config = translation_settings != {"inactive": "inactive"}
+    translation_lookup_path_in_pipeline_config = translation_lookup_path != "inactive"
+
+    if translation_settings_in_pipeline_config:
+
+        translation_directory = translation_settings.get("translation_directory", None)
+        translation_lookup_directory = translation_settings.get("translation_lookup_directory", None)
+        translation_backup_directory = translation_settings.get("translation_backup_directory", None)
+
+        new_translations_df = get_new_translations_from_completed_translations_directory(
+            translation_directory=translation_directory,
+            translation_lookup_path=translation_lookup_path,
+        )
+
+        if new_translations_df.count() > 0:
+
+            translation_lookup_df = extract_lookup_csv(translation_lookup_path, validation_schemas["csv_lookup_schema"])
+
+            new_translations_lookup_df = translation_lookup_df.union(new_translations_df)
+
+            backup_and_replace_translation_lookup_df(
+                new_lookup_df=new_translations_lookup_df,
+                lookup_directory=translation_lookup_directory,
+                backup_directory=translation_backup_directory,
+            )
+
+        df = translate_welsh_free_text_responses_digital(
+            df=df,
+            lookup_path=translation_lookup_path,
+        )
+
+        to_be_translated_df = get_welsh_responses_to_be_translated(df)
+
+        if to_be_translated_df.count() > 0:
+
+            export_responses_to_be_translated_to_translation_directory(
+                to_be_translated_df=to_be_translated_df, translation_directory=translation_directory
+            )
+
+    if translation_lookup_path_in_pipeline_config and not translation_settings_in_pipeline_config:
+        df = translate_welsh_free_text_responses_digital(
+            df=df,
+            lookup_path=translation_lookup_path,
+        )
+
+    df = translate_welsh_fixed_text_responses_digital(df)
+
     return df
