+++ cishouseholds/pipeline/translate.py
@@ -11,7 +11,6 @@ from cishouseholds.derive import translate_column_regex_replace
 from cishouseholds.edit import apply_value_map_multiple_columns
 from cishouseholds.edit import update_from_lookup_df
 from cishouseholds.hdfs_utils import rename
-from cishouseholds.pipeline.config import get_config
 from cishouseholds.pipeline.generate_outputs import write_csv_rename
 from cishouseholds.pipeline.input_file_processing import extract_lookup_csv
 from cishouseholds.pipeline.load import add_error_file_log_entry
@@ -127,6 +126,9 @@ def backup_and_replace_translation_lookup_df(
     """
     Backs up the lookup_path to the backup_directory with an automatically generated timestamp suffix.
     Writes the new_lookup_df to the lookup_directory
+
+    This uses two rename operations instead of a delete / write_csv_rename operations as there were obstinate
+    technical difficulties getting these to work as intended on HDFS
 
     Args:
         new_lookup_df (DataFrame): new_lookup_df to replace old_lookup_df
@@ -433,82 +435,3 @@ def get_welsh_responses_to_be_translated(df: DataFrame) -> DataFrame:
         .distinct()
     )
     return to_be_translated_df
-
-
-def translate_welsh_survey_responses_version_digital(df: DataFrame) -> DataFrame:
-    """
-    High-level function that manages the translation of both fixed-text and free-text Welsh responses in the
-    CIS Digital response extract.
-
-    Loads values from config to determine if the translation_project_workflow_enabled flag is true, if the
-    translation_project_workflow_enabled flag is true then it works through the translation project workflow to:
-    1. check for new completed free-text translations
-    2. if there are new completed free-text translations then backup and update the translations_lookup_df
-    3. update the df from the translations_lookup_df
-    4. filter the df to get free-text responses to be translated
-    5. export free-text responses to be translated
-
-    If the translation_project_workflow_enabled flag is false then it will just carry out step 3. above.
-    After this check it will update the fixed-text responses in the df.
-
-    This function requires that the config contains a valid translation_lookup_path in the storage dictionary of
-    the pipeline_config file.
-
-    Args:
-        df (DataFrame): df containing free- or fixed-text responses to translate
-
-    Returns:
-        DataFrame: df incorporating any available translations to free- or fixed-text responses
-    """
-    translation_settings = get_config().get("translation", {"inactive": "inactive"})
-    storage_settings = get_config().get("storage", {"inactive": "inactive"})
-    translation_lookup_path = storage_settings.get("translation_lookup_path", "inactive")
-
-    translation_settings_in_pipeline_config = translation_settings != {"inactive": "inactive"}
-    translation_lookup_path_in_pipeline_config = translation_lookup_path != "inactive"
-
-    if translation_settings_in_pipeline_config:
-
-        translation_directory = translation_settings.get("translation_directory", None)
-        translation_lookup_directory = translation_settings.get("translation_lookup_directory", None)
-        translation_backup_directory = translation_settings.get("translation_backup_directory", None)
-
-        new_translations_df = get_new_translations_from_completed_translations_directory(
-            translation_directory=translation_directory,
-            translation_lookup_path=translation_lookup_path,
-        )
-
-        if new_translations_df.distinct().count() > 0:
-
-            translation_lookup_df = extract_lookup_csv(translation_lookup_path, validation_schemas["csv_lookup_schema"])
-
-            new_translations_lookup_df = translation_lookup_df.union(new_translations_df)
-
-            backup_and_replace_translation_lookup_df(
-                new_lookup_df=new_translations_lookup_df,
-                lookup_directory=translation_lookup_directory,
-                backup_directory=translation_backup_directory,
-            )
-
-        df = translate_welsh_free_text_responses_digital(
-            df=df,
-            lookup_path=translation_lookup_path,
-        )
-
-        to_be_translated_df = get_welsh_responses_to_be_translated(df)
-
-        if to_be_translated_df.distinct().count() > 0:
-
-            export_responses_to_be_translated_to_translation_directory(
-                to_be_translated_df=to_be_translated_df, translation_directory=translation_directory
-            )
-
-    if translation_lookup_path_in_pipeline_config and not translation_settings_in_pipeline_config:
-        df = translate_welsh_free_text_responses_digital(
-            df=df,
-            lookup_path=translation_lookup_path,
-        )
-
-    df = translate_welsh_fixed_text_responses_digital(df)
-
-    return df
